+++ devmgr/app.py
@@ -1,6 +1,3 @@
-import logging
-import time
-
 import fxa.oauth
 import pyramid.renderers
 from pyramid.authorization import ACLAuthorizationPolicy
@@ -11,40 +8,7 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 
 from devmgr.db import Base
 from devmgr.resources import make_root
-from devmgr.security import DevMgrAuthenticationPolicy
-
-
-class Creds(object):
-    pass
-
-
-def verify_auth(event):
-    req = event.request
-    req.credentials = None
-    cred = Creds()
-    cred.recent = False
-    cred.account_id = None
-    token_str = req.headers.get("Authorization")
-    if not token_str:
-        return
-
-    d = token_str.strip().split()
-    if len(d) != 2:
-        logging.debug("Length of authorization is not 2")
-        return
-    typ, token = d
-    if typ != "Bearer":
-        logging.debug("Not a bearer token")
-        return
-
-    result = req.registry.fxa_oauth.verify_token(token)
-    created = result["created_at"] / 1000
-
-    cred.account_id = result["user"]
-    now = int(time.time())
-    if now-created < 600:
-        cred.recent = True
-    req.credentials = cred
+from devmgr.security import DevMgrAuthenticationPolicy, verify_auth
 
 
 def db(request):
